todo and users routes updated with auth

+++ TodoApplication/src/controller/todoController.js
@@ -1,6 +1,7 @@
 import {
   createTodoService,
   deleteTodoService,
+  getAllTodosService,
   getAllUsersTodosService,
   getTodoService,
   updateTodoService,
@@ -16,7 +17,13 @@ const handleTodoResponse = (res, status, message, size = 0, data = null) => {
 };
 const createTodo = async (req, res, next) => {
   try {
-    const todo = await createTodoService(req.body);
+    if (!req.user.id || !req.user.userName) {
+      return handleTodoResponse(res, 401, "Unauthorized! Please login.");
+    }
+    if (req.user.role == null) {
+      return handleTodoResponse(res, 403, "Unauthorized!");
+    }
+    const todo = await createTodoService(req.body, req.user);
     handleTodoResponse(
       res,
       201,
@@ -31,7 +38,20 @@ const createTodo = async (req, res, next) => {
 
 const getAllTodos = async (req, res, next) => {
   try {
-    const todo = await getAllUsersTodosService();
+    if (!req.user.id || !req.user.userName) {
+      return handleTodoResponse(res, 401, "Unauthorized! Please login.");
+    }
+    if (req.user.role !== "admin") {
+      const todo = await getAllUsersTodosService(req.user.id);
+      return handleTodoResponse(
+        res,
+        200,
+        "Todo fetched successfully!",
+        todo.size,
+        todo.data,
+      );
+    }
+    const todo = await getAllTodosService();
     handleTodoResponse(
       res,
       200,
@@ -46,7 +66,18 @@ const getAllTodos = async (req, res, next) => {
 
 const getTodoById = async (req, res, next) => {
   try {
+    if (!req.user.id || !req.user.userName) {
+      return handleTodoResponse(res, 401, "Unauthorized! Please login.");
+    }
     const todo = await getTodoService(req.params.id);
+    if (todo.data.user_id !== req.user.id && req.user.role !== "admin") {
+      return handleTodoResponse(
+        res,
+        403,
+        "Forbidden: You don't have access to this todo.",
+      );
+    }
+
     if (todo.size == 0) {
       return handleTodoResponse(res, 404, "Todo not found!");
     }
@@ -64,7 +95,17 @@ const getTodoById = async (req, res, next) => {
 
 const updateTodo = async (req, res, next) => {
   try {
+    if (!req.user.id || !req.user.userName) {
+      return handleTodoResponse(res, 401, "Unauthorized! Please login.");
+    }
     const todo = await updateTodoService(req.body, req.params.id);
+    if (todo.data.user_id !== req.user.id && req.user.role !== "admin") {
+      return handleTodoResponse(
+        res,
+        403,
+        "Forbidden: You don't have access to this todo.",
+      );
+    }
     if (todo.size == 0) {
       return handleTodoResponse(res, 404, "Todo not found!");
     }
@@ -82,7 +123,17 @@ const updateTodo = async (req, res, next) => {
 
 const deleteTodo = async (req, res, next) => {
   try {
+    if (!req.user.id || !req.user.userName) {
+      return handleTodoResponse(res, 401, "Unauthorized! Please login.");
+    }
     const todo = await deleteTodoService(req.params.id);
+    if (todo.data.user_id !== req.user.id && req.user.role !== "admin") {
+      return handleTodoResponse(
+        res,
+        403,
+        "Forbidden: You don't have access to this todo.",
+      );
+    }
     if (todo.size == 0) {
       return handleTodoResponse(res, 404, "Todo not found!");
     }
